always use volume mount option when mounting a partition
From ROS, if enable adb using the vol.up and vol.down buttons, the
/system partition is mounted by the function ensure_path_mounted() but
with hardcoded mount options.  As a consequence, the blocks are
modified and the reboot in MOS is blocked by the dm_verity feature
that detects a corruption.

This patch forces the function ensure_path_mounted() to use the mount
options from the volume structure, that were previously read from the
fstab.

Change-Id: I748d32c14cb821f4aae5bcc430089dab45375515