screen_ui: Allow setting screen margin space.
For round screen or screens with rounded corners, we don't want to show
texts within the margin which could otherwise be invisible.

Move the density computation to ScreenRecoveryUI ctor so that the value
can be used earlier.

Note that this is similar to the existing stuff in wear UI (outer_width,
outer_height). This CL gets ScreenRecoveryUI and WearRecoveryUI one-step
closer.

Bug: 62732748
Test: Setting and not setting margin_{width,height}_ on angler. Check the
      recovery texts (recovery menu as well as 'View recovery logs').
Change-Id: Ibf6238c9cc8949a42ed8a410e1c09d55b0b5a151