Expose EthernetDevice to custom UIs
Make it possible to build a custom UI which uses EthernetRecoveryUI.

Since EthernetRecoveryUI must be used with EthernetDevice, but the
make_device() function is the only place a custom UI can be specified,
we need to move the common EthernetDevice to the recovery_ui core and
make a new "default_ethernet_device" which looks like the other
default device library stubs, for backwards compatibility.

This makes it possible to use EthernetRecoveryUI on cuttlefish.

Bug: 172693524
Change-Id: I6a66bb949ef97d06f532adb286d2db4c002d2924